add move hooks
