+++ packages/good-design/src/hooks/index.ts
@@ -1,2 +1,1 @@
 export * from "./useQueryParam";
-export { default as useTimer } from "./useTimer";
