+++ packages/sdk-v2/src/sdk/base/index.ts
@@ -1,2 +1,4 @@
 export * from './react';
 export * from './sdk';
+export * from './utils';
+export * from './hooks';
