+++ packages/sdk-v2/src/sdk/base/utils/linking.ts
